Factor out card component, move to home page section organization
