+++ src/Basics/HomePageSection.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Styles from './Styles';
-import Paper from '@material-ui/core/Paper';
 import HideIcon from '@material-ui/icons/VisibilityOff';
 import { useTranslation } from 'react-i18next';
+import HomePageCard from '../Components/Patient/HomePageCard';
 
-const HomePageCard = (props) => {
+
+const HomePageSection = (props) => {
 
     const classes = useStyles();
     const { t } = useTranslation('translation');
@@ -16,9 +17,7 @@ const HomePageCard = (props) => {
 
     return (<div className={classes.superContainer} >
         {props.upperText && <span className={classes.upperText}>{props.upperText}</span>}
-        <Paper id={props.id} className={`${classes.container}  ${props.noPadding && classes.noPadding} ${props.className}`}>
-            {props.children}
-        </Paper>
+        <HomePageCard {...props} />
         {props.hideCard && <div className={classes.bottomText} onClick={props.hideCard}>
             <span>{t('commonWords.hideThis')}</span>
             <HideIcon style={{ fontSize: "1.25em", marginLeft: ".5em" }} />
@@ -29,17 +28,8 @@ const HomePageCard = (props) => {
 const useStyles = makeStyles({
 
     superContainer: {
-        width: "100vw",
+        width: "100%",
 
-    },
-    container: {
-        ...Styles.modifiedPaper,
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        marginBottom: "1em",
-        padding: ".5em",
-        paddingTop: "1em",
     },
     upperText: {
         ...Styles.secondaryText,
@@ -64,4 +54,4 @@ const useStyles = makeStyles({
 })
 
 
-export default HomePageCard;
+export default HomePageSection;
