+++ src/Patient/Home/ActionBox.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import NewButton from '../../Basics/NewButton';
 import Clipboard from '@material-ui/icons/Assignment';
 import Camera from '@material-ui/icons/CameraAlt';
-import InteractionCard from '../../Basics/HomePageCard';
+import InteractionCard from '../../Basics/HomePageSection';
 import useStores from '../../Basics/UseStores';
 import { observer } from 'mobx-react'
 import { useTranslation } from 'react-i18next';
