+++ src/Patient/Home/SimpleProgressCard.js
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react'
 import useStores from '../../Basics/UseStores'
 import { makeStyles } from '@material-ui/core/styles'
 import Styles from '../../Basics/Styles';
-import InteractionCard from '../../Basics/HomePageCard';
+import InteractionCard from '../../Basics/HomePageSection';
 import Colors from '../../Basics/Colors';
 import { useTranslation } from 'react-i18next';
 import ClickableText from '../../Basics/ClickableText';
